refactor: hook
