+++ src/frontend/web/src/hook/updateChannelStore.ts
@@ -1,0 +1,13 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { Params } from 'react-router-dom';
+import { selectChannel } from '../modules/channel';
+
+export default function updateChannelStoreHook(params: Params) {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const { channelId } = params;
+    if (channelId) dispatch(selectChannel(channelId));
+  }, [params]);
+}
